Structured project to support multiple platforms.
