+++ desktop/swing/Main.java
@@ -1,0 +1,8 @@
+package edu.lasalle.gdi.convocatoria.sopar.desktop.swing;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
